+++ src/data/trade.rs
@@ -1,6 +1,6 @@
 //! The trade-related operations.
 
-use std::io::{BufRead, BufReader};
+use std::io::BufReader;
 
 use crypto_msg_parser::TradeMsg;
 use rust_decimal::prelude::ToPrimitive;
@@ -54,8 +54,7 @@ pub fn decode_trade(payload: &[u8]) -> TradeResult<TradeMsg> {
     let exchange_timestamp = ExchangeTimestampRepr::try_from_reader(&mut reader)?.0;
 
     // 2. 收到时间戳: 6 字节时间戳 (NOT USED)
-    reader.consume(8);
-    // let received_timestamp = ReceivedTimestampRepr::try_from_reader(&mut reader)?;
+    ReceivedTimestampRepr::try_from_reader(&mut reader)?;
 
     // 3. EXCHANGE: 1 字节信息标识
     let exchange_type = ExchangeTypeRepr::try_from_reader(&mut reader)?.0;